requirements.txt
pytest
requests
sqlalchemy
factory_boy
pymysql
